Update spec status layout
