+++ src/Main/Layout/Footer.css
@@ -1,6 +1,6 @@
 footer {
   border-top: 2px solid rgba(0, 0, 0, 0.9);
-  background: rgba(18, 27, 31, 0.85);
+  background: rgba(20, 26, 29, 0.85);
   padding-top: 30px;
   padding-bottom: 90px;
 }
