+++ src/Main/SpecListing.css
@@ -1,0 +1,26 @@
+.spec-card {
+  background: rgba(9, 20, 27, 0.7);
+  border: 1px solid #000;
+  border-radius: 6px;
+  overflow: hidden;
+  box-shadow: 0 6px 13px 0 rgba(0,0,0,.2), 0 0 2px rgba(0,0,0,.5);
+}
+.spec-card .icon {
+  width: 140px;
+  height: 140px;
+  background-repeat: no-repeat;
+  background-size: cover;
+  border-right: 1px solid #000;
+}
+.spec-card .description {
+  padding: 12px 15px;
+}
+.spec-card .description h1 {
+  margin-bottom: 10px;
+}
+.spec-card .maintainer-name {
+  background: #b24747;
+  padding: 3px 4px;
+  margin: -3px 0;
+  border-radius: 3px;
+}
